advertisement/requirements.txt
asgiref==3.4.1
Django==3.2.9
pytz==2021.3
sqlparse==0.4.2
typing-extensions==4.0.1
psycopg2
djangorestframework==3.12.4
requests